+++ thejoker/prior_helpers.py
@@ -1,103 +1,15 @@
 # Third-party
 import astropy.units as u
-import numpy as np
-import pymc3 as pm
-from pymc3.distributions import generate_samples
-import theano.tensor as tt
-import exoplanet.units as xu
 
-__all__ = ['UniformLog', 'FixedCompanionMass']
-
-
-class UniformLog(pm.Continuous):
-
-    def __init__(self, a, b, **kwargs):
-        """A distribution over a value, x, that is uniform in log(x) over the
-        domain :math:`(a, b)`.
-        """
-
-        self.a = float(a)
-        self.b = float(b)
-        assert (self.a > 0) and (self.b > 0)
-        self._fac = np.log(self.b) - np.log(self.a)
-
-        shape = kwargs.get("shape", None)
-        if shape is None:
-            testval = 0.5 * (self.a + self.b)
-        else:
-            testval = 0.5 * (self.a + self.b) + np.zeros(shape)
-        kwargs["testval"] = kwargs.pop("testval", testval)
-        super(UniformLog, self).__init__(**kwargs)
-
-    def _random(self, size=None):
-        uu = np.random.uniform(size=size)
-        return np.exp(uu * self._fac + np.log(self.a))
-
-    def random(self, point=None, size=None):
-        return generate_samples(
-            self._random,
-            dist_shape=self.shape,
-            broadcast_shape=self.shape,
-            size=size,
-        )
-
-    def logp(self, value):
-        return tt.ones_like(value) * -np.log(self._fac)
+__all__ = ['get_nonlinear_equiv_units',
+           'validate_poly_trend',
+           'get_linear_equiv_units',
+           'validate_n_offsets',
+           'get_v0_offsets_equiv_units']
 
 
-Kprior_str = r"p(K) \propto \mathcal{N}(K \,|\, \mu_K, \sigma_K)"
-sigmaK_str = (r"\sigma_K = \sigma_{K, 0} \, \left(\frac{P}{P_0}\right)^{-1/3} "
-              r"\, \left(1 - e^2\right)^{-1}")
-
-
-class FixedCompanionMass(pm.Normal):
-    f"""
-    A distribution over velocity semi-amplitude, :math:`K`, that, at
-    fixed primary mass, is a fixed Normal distribution in companion mass. This
-    has the form:
-
-    .. math::
-
-        {Kprior_str}
-        {sigmaK_str}
-
-    where :math:`P` and :math:`e` are period and eccentricity, and
-    ``sigma_K0`` and ``P0`` are parameters of this distribution that must
-    be specified.
-    """
-
-    @u.quantity_input(sigma_K0=u.km/u.s, P0=u.day, max_K=u.km/u.s)
-    def __init__(self, P, e, sigma_K0, P0, mu=0., max_K=500*u.km/u.s,
-                 K_unit=None, **kwargs):
-        self._sigma_K0 = sigma_K0
-        self._P0 = P0
-        self._max_K = max_K
-
-        if K_unit is not None:
-            self._sigma_K0 = self.sigma_K0.to(K_unit)
-        self._max_K = self._max_K.to(self._sigma_K0.unit)
-
-        if hasattr(P, xu.UNIT_ATTR_NAME):
-            self._P0 = self._P0.to(getattr(P, xu.UNIT_ATTR_NAME))
-
-        sigma_K0 = self._sigma_K0.value
-        P0 = self._P0.value
-
-        sigma = tt.min([self._max_K.value,
-                        sigma_K0 * (P/P0)**(-1/3) / np.sqrt(1-e**2)])
-        super().__init__(mu=mu, sigma=sigma)
-
-
-def _get_v_trend_names(poly_trend):
-    return ['v{0}'.format(i) for i in range(poly_trend)]
-
-
-def _validate_polytrend(poly_trend):
-    poly_trend = int(poly_trend)
-    return poly_trend, _get_v_trend_names(poly_trend)
-
-
-def _get_nonlinear_equiv_units():
+# Nonlinear parameter helpers:
+def get_nonlinear_equiv_units():
     return {
         'P': u.day,
         'e': u.one,
@@ -107,8 +19,74 @@ def _get_nonlinear_equiv_units():
     }
 
 
-def _get_linear_equiv_units(v_names):
+# Linear parameter helpers:
+def validate_poly_trend(poly_trend):
+    try:
+        poly_trend = int(poly_trend)
+    except Exception:
+        raise ValueError("poly_trend must be an integer that specifies the "
+                         "number of polynomial (in time) trend terms to "
+                         "include in The Joker.")
+    vtrend_names = ['v{0}'.format(i) for i in range(poly_trend)]
+    return poly_trend, vtrend_names
+
+
+def get_linear_equiv_units(poly_trend):
+    poly_trend, v_names = validate_poly_trend(poly_trend)
     return {
         'K': u.m/u.s,
         **{name: u.m/u.s/u.day**i for i, name in enumerate(v_names)}
     }
+
+
+def validate_sigma_v(sigma_v, poly_trend, v_names):
+    if isinstance(sigma_v, u.Quantity):
+        if not sigma_v.isscalar:
+            raise ValueError("You must pass in a scalar value for sigma_v if "
+                             "passing in a single quantity.")
+        sigma_v = {'v0': sigma_v}
+
+    if hasattr(sigma_v, 'keys'):
+        for name in v_names:
+            if name not in sigma_v.keys():
+                raise ValueError("If specifying the standard-deviations of "
+                                 "the polynomial trend parameter prior, you "
+                                 "must pass in values for all parameter names."
+                                 "Expected keys: {}, received: {}"
+                                 .format(v_names, sigma_v.keys()))
+        return sigma_v
+
+    try:
+        if len(sigma_v) != poly_trend:
+            raise ValueError("You must pass in a single sigma value for "
+                             "each velocity trend parameter: You passed in "
+                             "{} values, but poly_trend={}"
+                             .format(len(sigma_v), poly_trend))
+        sigma_v = {name: val for name, val in zip(v_names, sigma_v)}
+
+    except TypeError:
+        raise TypeError("Invalid input for velocity trend prior sigma "
+                        "values. This must either be a scalar Quantity (if "
+                        "poly_trend=1) or an iterable of Quantity objects "
+                        "(if poly_trend>1)")
+
+    return sigma_v
+
+
+# Offsets in v0 helpers:
+def validate_n_offsets(n_offsets):
+    try:
+        n_offsets = int(n_offsets)
+    except Exception:
+        raise ValueError("n_offsets must be an integer that specifies the "
+                         "number of v0 offset parameters to include in "
+                         "The Joker. These parameters allow passing in data "
+                         "from multiple surveys that may have unknown "
+                         "calibration offsets.")
+    offset_names = ['dv0_{0}'.format(i) for i in range(1, n_offsets+1)]
+    return n_offsets, offset_names
+
+
+def get_v0_offsets_equiv_units(n_offsets):
+    n_offsets, names = validate_n_offsets(n_offsets)
+    return {name: u.m/u.s for i, name in enumerate(names)}
